Move calling Update in transformation to separate class
